fast-fix

+++ src/main/java/by/shaaldy/booking/entity/Hotel.java
@@ -1,5 +1,6 @@
 package by.shaaldy.booking.entity;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -31,13 +32,12 @@ public class Hotel {
   @Column(nullable = false, length = 100)
   private String city;
 
-  @Column
-  private String address;
+  @Column private String address;
 
   @DecimalMin(value = "0.0", inclusive = false, message = "Rating must be greater than 0")
   @DecimalMax(value = "5.0", message = "Rating cannot exceed 5")
   @Column(precision = 2, scale = 1)
-  private Double rating;
+  private BigDecimal rating;
 
   @Column(name = "created_at", nullable = false, updatable = false)
   private LocalDateTime createdAt;
